pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ebbingcontext"
version = "0.1.0"
description = "Ebbinghaus Forgetting Curve-Based Context Management Engine for LLM Agents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "wang lian", email = "liainseptember@gmail.com" },
]
keywords = ["llm", "agent", "memory", "context", "mcp", "ebbinghaus"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "mcp>=1.0.0",
    "chromadb>=0.4.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
pgvector = ["psycopg2-binary>=2.9.0", "pgvector>=0.2.0"]
bge = ["sentence-transformers>=2.2.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
ebbingcontext = "ebbingcontext.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100
